Define IAsyncStorage abstraction and use it to generalize ExtensionLocalStorage
